Implementacion de la base de datos, la clase Editar y la clase Ver. La aplicacion ya tiene funcionalidad.

+++ app/src/main/java/com/grupo5/todo_ocio/list/Lugar.java
@@ -56,11 +56,11 @@ public class Lugar {
 
     public double getLongitud() { return longitud; }
 
-    public void setLongitud(float longitud) { this.longitud = longitud; }
+    public void setLongitud(double longitud) { this.longitud = longitud; }
 
     public double getLatitud() { return latitud; }
 
-    public void setLatitud(float latitud) { this.latitud = latitud; }
+    public void setLatitud(double latitud) { this.latitud = latitud; }
 
     public String getImagen() {return imagen; }
